Move shared import arguments to parent class (#645)
* Move shared args to parent class.

* Add to collection args.

* Promote check_healpix_order_range

* Use hats branch

* Move up arguments to create additional metadata files.

* Add more common args from collection.
