+++ src/hats_import/collection/arguments.py
@@ -208,6 +208,14 @@ class CollectionArguments(RuntimeArguments):
         useful_kwargs = {
             "output_path": self.catalog_path,
             "addl_hats_properties": self.addl_hats_properties,
+            "npix_suffix": self.npix_suffix,
+            "npix_parquet_name": self.npix_parquet_name,
+            "write_table_kwargs": self.write_table_kwargs,
+            "should_write_skymap": self.should_write_skymap,
+            "row_group_kwargs": self.row_group_kwargs,
+            "skymap_alt_orders": self.skymap_alt_orders,
+            "create_thumbnail": self.create_thumbnail,
+            "create_metadata": self.create_metadata,
             "tmp_dir": self.tmp_path,
             "resume": self.resume,
             "progress_bar": self.progress_bar,
