+++ tests/hats_import/margin_cache/test_margin_cache.py
@@ -4,6 +4,7 @@ import os
 
 import numpy.testing as npt
 import pandas as pd
+import pyarrow.parquet as pq
 import pytest
 from hats import read_hats
 from hats.io import get_parquet_metadata_pointer, paths
@@ -29,6 +30,7 @@ def test_margin_cache_gen(small_sky_source_catalog, tmp_path, dask_client):
         output_artifact_name="catalog_cache",
         margin_order=8,
         progress_bar=False,
+        npix_suffix=".pq",
     )
 
     assert args.catalog.catalog_info.ra_column == "source_ra"
@@ -38,7 +40,7 @@ def test_margin_cache_gen(small_sky_source_catalog, tmp_path, dask_client):
     norder = 1
     npix = 47
 
-    test_file = paths.pixel_catalog_file(args.catalog_path, HealpixPixel(norder, npix))
+    test_file = paths.pixel_catalog_file(args.catalog_path, HealpixPixel(norder, npix), npix_suffix=".pq")
 
     data = pd.read_parquet(test_file)
 
@@ -63,10 +65,15 @@ def test_margin_cache_gen(small_sky_source_catalog, tmp_path, dask_client):
     catalog = read_hats(args.catalog_path)
     assert catalog.on_disk
     assert catalog.catalog_path == args.catalog_path
+    assert catalog.catalog_info.npix_suffix == ".pq"
 
     # Check that the data thumbnail does not exist. It should only exist for
     # main object/source catalogs.
     assert not (args.catalog_path / "dataset" / "data_thumbnail.parquet").exists()
+
+    metadata = pq.read_metadata(test_file)
+    assert metadata.num_row_groups == 1
+    assert metadata.row_group(0).column(0).compression == "ZSTD"
 
 
 @pytest.mark.dask(timeout=150)
